test identifier evaluation
to test with arbitrary environment, move the test code inside.
